requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
jwt==1.3.1
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.1
